made a change

+++ app/Models/User.php
@@ -192,6 +192,6 @@ class User extends Authenticatable
     }
     public function refPercent($CPV)
     {
-        return (100 * $this->cpv)/$CPV;
+        return (int)(100 * $this->cpv)/$CPV;
     }
 }
